refactor: split adapter generators into separate packages
Core EquatableGenerator no longer knows about DataContract or MessagePack.
RegisterProvider is now public static so adapter generators can call it directly.
GetBaseEquatableType detects any *EquatableAttribute in Equatable.Attributes
namespace instead of hard-coding adapter names.

New packages:
- Equatable.SourceGenerator.DataContract — DataContractEquatableGenerator
- Equatable.SourceGenerator.MessagePack — MessagePackEquatableGenerator
- Equatable.Generator.DataContract — NuGet wrapper + DataContractEquatableAttribute
- Equatable.Generator.MessagePack — NuGet wrapper + MessagePackEquatableAttribute

DataContractEquatableAttribute and MessagePackEquatableAttribute moved out of
Equatable.Generator into their respective adapter packages.
