+++ src/Equatable.Generator.DataContract/Attributes/DataContractEquatableAttribute.cs
