+++ src/Equatable.Generator.DataContract/Equatable.Generator.DataContract.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard2.0;net8.0;net9.0;net10.0</TargetFrameworks>
+    <RootNamespace>Equatable</RootNamespace>
+    <Description>Source generator for Equals and GetHashCode for types using System.Runtime.Serialization DataMember attributes</Description>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Include="..\Equatable.SourceGenerator\bin\$(Configuration)\netstandard2.0\Equatable.SourceGenerator.dll" PackagePath="analyzers/dotnet/roslyn4.8/cs" Pack="true" Visible="false" />
+    <None Include="..\Equatable.SourceGenerator.DataContract\bin\$(Configuration)\netstandard2.0\Equatable.SourceGenerator.DataContract.dll" PackagePath="analyzers/dotnet/roslyn4.8/cs" Pack="true" Visible="false" />
+    <None Include="..\Equatable.SourceGenerator\Equatable.Generator.targets" PackagePath="build" Pack="true" Visible="false" />
+  </ItemGroup>
+
+</Project>
