+++ src/Equatable.Generator.MessagePack/Equatable.Generator.MessagePack.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard2.0;net8.0;net9.0;net10.0</TargetFrameworks>
+    <RootNamespace>Equatable</RootNamespace>
+    <Description>Source generator for Equals and GetHashCode for types using MessagePack Key attributes</Description>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Include="..\Equatable.SourceGenerator\bin\$(Configuration)\netstandard2.0\Equatable.SourceGenerator.dll" PackagePath="analyzers/dotnet/roslyn4.8/cs" Pack="true" Visible="false" />
+    <None Include="..\Equatable.SourceGenerator.MessagePack\bin\$(Configuration)\netstandard2.0\Equatable.SourceGenerator.MessagePack.dll" PackagePath="analyzers/dotnet/roslyn4.8/cs" Pack="true" Visible="false" />
+    <None Include="..\Equatable.SourceGenerator\Equatable.Generator.targets" PackagePath="build" Pack="true" Visible="false" />
+  </ItemGroup>
+
+</Project>
