+++ src/Equatable.SourceGenerator.DataContract/DataContractEquatableGenerator.cs
@@ -1,0 +1,39 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Equatable.SourceGenerator.DataContract;
+
+[Generator]
+public class DataContractEquatableGenerator : IIncrementalGenerator
+{
+    public void Initialize(IncrementalGeneratorInitializationContext context)
+    {
+        EquatableGenerator.RegisterProvider(context,
+            fullyQualifiedMetadataName: "Equatable.Attributes.DataContractEquatableAttribute",
+            trackingName: "DataContractEquatableAttribute",
+            propertyFilter: IsIncludedDataContract);
+    }
+
+    private static bool IsIncludedDataContract(IPropertySymbol propertySymbol)
+    {
+        if (propertySymbol.IsIndexer || propertySymbol.DeclaredAccessibility != Accessibility.Public)
+            return false;
+
+        var attributes = propertySymbol.GetAttributes();
+        if (attributes.Length == 0)
+            return false;
+
+        if (attributes.Any(a => a.AttributeClass is
+            {
+                Name: "IgnoreDataMemberAttribute",
+                ContainingNamespace: { Name: "Serialization", ContainingNamespace: { Name: "Runtime", ContainingNamespace.Name: "System" } }
+            }))
+            return false;
+
+        return attributes.Any(a => a.AttributeClass is
+        {
+            Name: "DataMemberAttribute",
+            ContainingNamespace: { Name: "Serialization", ContainingNamespace: { Name: "Runtime", ContainingNamespace.Name: "System" } }
+        });
+    }
+}
